newly added pack - getX new page added - orderContainer get the total cart amount set the data to firestore retrieve the data from order collection item page - small alignment to checkout gestures button

+++ lib/itemPage.dart
@@ -434,7 +434,7 @@ class _ItemPageState extends State<ItemPage> {
                         'downloadurl': widget.data['downloadurl'],
                         'net': widget.data['net'],
                         'symbol': widget.data['symbol'],
-                        'price': double.parse(widget.data['price'])*counter,
+                        'price': int.parse(widget.data['price'])*counter,
                         'quantity':counter
                       }
                   );
